+++ src/Imaging/AbstractImageFilter.h
@@ -55,8 +55,8 @@ protected:
 
 	void
 	SetOutputImageSize( 
-			size_t 		minimums[], 
-			size_t 		maximums[], 
+			int32 		minimums[], 
+			int32 		maximums[], 
 			float32		elementExtents[]
 		    );
